New option: --verbose for detailed output
Now in normal mode, the tests that passed are not shown in the output,
only those that have failed are shown. The idea is that usually people
want to see that all tests have passed (the OK summary line at the end),
otherwise, which were the failed ones (with details).

So, the normal output now when all the test pass is just the summary
line. Example:

    > ./doctest.fish tests/status.md
    tests/status.md: OK (5 tests passed)

Using the new `--verbose` option, tests that passed can be shown again.

    > ./doctest.fish --verbose tests/status.md
    tests/status.md:3: [ ok ] true; echo $status
    tests/status.md:5: [ ok ] false; echo $status
    tests/status.md:7: [ ok ] echo "command output and status"; echo $status
    tests/status.md:13: [ ok ] false
    tests/status.md:14: [ ok ] echo $status  # DOES NOT WORK
    tests/status.md: OK (5 tests passed)

Since this option conflicts with `--quiet`, it's not possible to use
both at the same time.