Refactor: separate parser and test runner
Previous code was reading, parsing and running the tests for each file,
on a line by line basis: read a new line, parse it, execute the tests if
now it's the right time to do it.

Now there's two major steps:

- First parse the whole input file looking for the test cases. Save the
  results to a global `$parsed_data` variable. This is an internal
  representation of the test cases, a list where each item is a
  `line_number:type:contents` record. Example:

      3:cmd:true; echo $status
      4:out:0
      5:cmd:false; echo $status
      6:out:1
      7:cmd:echo "command output and status"; echo $status
      8:out:command output and status
      9:out:0

- The second step is to iterate over `$parsed_data` and run the tests
  defined there.

The functions to handle those steps are:

- parse_input_file
- test_input_file

Note that the very last task of showing the summary line for the file
was moved to a new `show_file_summary` function.

For the future, this internal representation of all the tests can be
easily converted to different formats and be dumped to the user, so
other tools can act on that data.