Fix: Prefix variables and functions with __
This "uglify" change fixes a name clash bug.

Since tests are run with `eval`, they can interfere with `doctest.fish`
own execution. For example, when a user test defines a variable
`$failed_tests`, or a function named `error`, they are directly
affecting the tester code.

To help mitigate that name clash, all the variables and functions now
have a `__` prefix. The exceptions are local variables that are isolated
into functions and are safe to be left unprefixed.